requirements.txt
flask
pymysql
gunicorn
python-dotenv
openai 
oss2 
Pillow
Frozen-Flask